refactor : step1 피드백 반영
1. Excpetion 위치 변경
2. Line 중복체크 메서드 명 변경
3. Line 목록 validation 제거 및 테스트 제거
