+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -6,8 +6,8 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import nextstep.subway.exception.LineDuplicateException;
-import nextstep.subway.exception.NotFoundLineException;
+import nextstep.subway.line.exception.LineDuplicateException;
+import nextstep.subway.line.exception.NotFoundLineException;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
 import nextstep.subway.line.dto.LineRequest;
@@ -16,7 +16,6 @@ import nextstep.subway.line.dto.LineResponse;
 @Service
 @Transactional
 public class LineService {
-    public static final int LINES_EMPTY = 0;
     private LineRepository lineRepository;
 
     public LineService(LineRepository lineRepository) {
@@ -24,12 +23,12 @@ public class LineService {
     }
 
     public LineResponse saveLine(LineRequest request) {
-        saveLineValidation(request);
+        checkDuplicateLine(request);
         Line line = lineRepository.save(request.toLine());
         return LineResponse.of(line);
     }
 
-    private void saveLineValidation(LineRequest request) {
+    private void checkDuplicateLine(LineRequest request) {
         if (isDuplicate(request.getName())) {
             throw new LineDuplicateException();
         }
@@ -42,20 +41,9 @@ public class LineService {
     @Transactional()
     public List<LineResponse> findAllLines() {
         List<Line> lines = lineRepository.findAll();
-        findAllLinesValidation(lines);
         return lines.stream()
             .map(LineResponse::of)
             .collect(Collectors.toList());
-    }
-
-    private void findAllLinesValidation(List<Line> lines) {
-        if (isLinesEmpty(lines)) {
-            throw new NotFoundLineException();
-        }
-    }
-
-    private boolean isLinesEmpty(List<Line> lines) {
-        return lines.size() <= LINES_EMPTY;
     }
 
     @Transactional(readOnly = true)
