+++ src/main/java/nextstep/subway/line/exception/LineDuplicateException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.subway.line.exception;
 
 public class LineDuplicateException extends RuntimeException {
     public static final String LINE_DUPLICATE = "이미 생성된 노선입니다.";
